refactor(components): extract collateral preview helpers
