+++ frontend/src/components/CollateralPreview.tsx
@@ -1,17 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import { ArrowRight, AlertTriangle, Coins, RefreshCw } from 'lucide-react';
 import { BitflowSDK } from '@bitflowlabs/core-sdk';
-import { formatPercentage, formatSTX, formatUSD } from '../utils/formatters';
-
-type BitflowQuoteResult = Awaited<ReturnType<BitflowSDK['getQuoteForRoute']>>;
-
-type PreviewRoute = NonNullable<BitflowQuoteResult['bestRoute']> & {
-  tokenYAmount?: number | null;
-  priceImpact?: number | null;
-  priceImpactPercent?: number | null;
-  price_impact?: number | null;
-  impact?: number | null;
-};
+import { formatSTX, formatUSD } from '../utils/formatters';
+import {
+  extractEstimatedOutput,
+  extractPriceImpact,
+  formatPriceImpact,
+  getRouteLabel,
+  normalizeAmount,
+  type BitflowQuoteResult,
+  type PreviewRoute,
+} from './collateralPreviewUtils';
 
 type PreviewQuoteResult = Omit<BitflowQuoteResult, 'bestRoute'> & {
   bestRoute: PreviewRoute | null;
@@ -28,112 +27,6 @@ const bitflow = new BitflowSDK();
 const DEBOUNCE_MS = 500;
 const INPUT_TOKEN = 'token-stx';
 const OUTPUT_TOKEN = 'token-usda';
-
-const TOKEN_LABELS: Record<string, string> = {
-  'token-stx': 'STX',
-  'token-usda': 'USDA',
-};
-
-const normalizeAmount = (value: number): number | null => {
-  if (!Number.isFinite(value) || value <= 0) {
-    return null;
-  }
-
-  return value;
-};
-
-const formatTokenLabel = (token: string): string => {
-  return TOKEN_LABELS[token] ?? token.replace(/[-_]/g, ' ').toUpperCase();
-};
-
-const formatDexLabel = (dex: string): string => {
-  return dex.replace(/[-_]/g, ' ').toUpperCase();
-};
-
-const extractEstimatedOutput = (bestRoute: PreviewRoute | null): number | null => {
-  if (!bestRoute) {
-    return null;
-  }
-
-  const candidate = bestRoute.tokenYAmount ?? bestRoute.quote;
-
-  return typeof candidate === 'number' && Number.isFinite(candidate) && candidate > 0
-    ? candidate
-    : null;
-};
-
-const readNumericValue = (...values: Array<unknown>): number | null => {
-  for (const value of values) {
-    if (typeof value === 'number' && Number.isFinite(value)) {
-      return value;
-    }
-
-    if (typeof value === 'string' && value.trim().length > 0) {
-      const parsed = Number(value);
-
-      if (Number.isFinite(parsed)) {
-        return parsed;
-      }
-    }
-  }
-
-  return null;
-};
-
-const extractPriceImpact = (bestRoute: PreviewRoute | null): number | null => {
-  if (!bestRoute) {
-    return null;
-  }
-
-  const params = bestRoute.params as
-    | { priceImpact?: unknown; price_impact?: unknown; 'price-impact'?: unknown }
-    | undefined;
-  const quoteParams = bestRoute.quoteData?.parameters as
-    | { priceImpact?: unknown; price_impact?: unknown; 'price-impact'?: unknown }
-    | undefined;
-
-  const rawValue = readNumericValue(
-    bestRoute.priceImpact,
-    bestRoute.priceImpactPercent,
-    bestRoute.price_impact,
-    bestRoute.impact,
-    params?.priceImpact,
-    params?.price_impact,
-    params?.['price-impact'],
-    quoteParams?.priceImpact,
-    quoteParams?.price_impact,
-    quoteParams?.['price-impact']
-  );
-
-  if (rawValue === null) {
-    return null;
-  }
-
-  if (Math.abs(rawValue) <= 1) {
-    return rawValue * 100;
-  }
-
-  return rawValue;
-};
-
-const formatPriceImpact = (priceImpact: number): string => {
-  return formatPercentage(Math.abs(priceImpact), 2);
-};
-
-const getRouteLabel = (bestRoute: PreviewRoute | null): string => {
-  if (!bestRoute) {
-    return 'Live Bitflow route';
-  }
-
-  const tokenPath = bestRoute.tokenPath.length > 0
-    ? bestRoute.tokenPath.map(formatTokenLabel).join(' → ')
-    : 'STX → USDA';
-  const dexPath = bestRoute.dexPath.length > 0
-    ? bestRoute.dexPath.map(formatDexLabel).join(' / ')
-    : 'Bitflow';
-
-  return `${tokenPath} via ${dexPath}`;
-};
 
 /**
  * CollateralPreview Component
